src/requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
openai==1.3.4
pycryptodome==3.17
python-dotenv==1.0.0
